# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go API Redis",
        "title": "Go API Redis",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/order": {
            "post": {
                "description": "Set order",
                "tags": [
                    "order"
                ],
                "summary": "Set",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "description": "Get order",
                "tags": [
                    "order"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "order.Order": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Delivered"
                },
                "key": {
                    "type": "string",
                    "example": "7e1efed0-e86a-4286-8c4a-a609751f9a0b"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "updated": {
                    "type": "string",
                    "example": "2022-04-17T12:25:43.511Z"
                }
            }
        }
    }
}